+++ resources/views/admin/pages/notifications/index.blade.php
@@ -104,7 +104,9 @@
                                                     <th>Tiêu đề</th>
                                                     <th>Kiểu</th>
                                                     <th>Nội dung</th>
-                                                    <th>Hành động</th>
+                                                    @if (Auth()->user()->can('update', \App\Models\Notification::class) || Auth()->user()->can('delete', \App\Models\Notification::class))
+                                                        <th>Hành động</th>
+                                                    @endif
                                                 </tr>
                                             </thead>
                                             <tbody>
@@ -134,35 +136,35 @@
                                                                 Tin khuyến mãi
                                                             @endif
                                                         </td>
-
                                                         <td>
                                                             {!! $notification->content !!}
                                                         </td>
-
-                                                        <td>
-                                                            <div
-                                                                style="padding-right: 20px; display: flex; justify-content: end">
-                                                                @can('update', \App\Models\Notification::class)
-                                                                    <a href="{{ route('admin.notifications.edit', $notification->id) }}"
-                                                                        class="btn btn-primary shadow btn-xs sharp me-1">
-                                                                        <i class="fa fa-pencil"></i>
-                                                                    </a>
-                                                                @endcan
-                                                                @can('delete', \App\Models\Notification::class)
-                                                                    <form
-                                                                        action="{{ route('admin.notifications.delete', $notification->id) }}"
-                                                                        class="formDelete" method="post">
-                                                                        @csrf
-                                                                        @method('DELETE')
-                                                                        <button
-                                                                            class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
-                                                                            data-type="DELETE" href="">
-                                                                            <i class="fa fa-trash"></i>
-                                                                        </button>
-                                                                    </form>
-                                                                @endcan
-                                                            </div>
-                                                        </td>
+                                                        @if (Auth()->user()->can('update', \App\Models\Notification::class) || Auth()->user()->can('delete', \App\Models\Notification::class))
+                                                            <td>
+                                                                <div
+                                                                    style="padding-right: 20px; display: flex; justify-content: end">
+                                                                    @can('update', \App\Models\Notification::class)
+                                                                        <a href="{{ route('admin.notifications.edit', $notification->id) }}"
+                                                                            class="btn btn-primary shadow btn-xs sharp me-1">
+                                                                            <i class="fa fa-pencil"></i>
+                                                                        </a>
+                                                                    @endcan
+                                                                    @can('delete', \App\Models\Notification::class)
+                                                                        <form
+                                                                            action="{{ route('admin.notifications.delete', $notification->id) }}"
+                                                                            class="formDelete" method="post">
+                                                                            @csrf
+                                                                            @method('DELETE')
+                                                                            <button
+                                                                                class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
+                                                                                data-type="DELETE" href="">
+                                                                                <i class="fa fa-trash"></i>
+                                                                            </button>
+                                                                        </form>
+                                                                    @endcan
+                                                                </div>
+                                                            </td>
+                                                        @endif
                                                     </tr>
                                                 @endforeach
                                             </tbody>
@@ -177,7 +179,7 @@
                                     <div class="d-flex justify-content-center align-items-center p-5">
                                         <div>
                                             <h3 class="text-center">
-                                                Không tìm thấy: {{ request()->title }}
+                                                {{ request()->name ? 'Không có kết quả với từ khóa:' . request()->name : 'Không có dữ liệu' }}
                                             </h3>
                                         </div>
                                     </div>
