+++ resources/views/admin/pages/rewards/index.blade.php
@@ -119,8 +119,10 @@
                                                     <th>Tên</th>
                                                     <th>Điểm</th>
                                                     <th>Hình ảnh</th>
-                                                    <th>Trạng thái</th>
-                                                    <th>Hành động</th>
+                                                    <th class="text-center">Trạng thái</th>
+                                                    @if (auth()->user()->can('update', \App\Models\Reward::class) || auth()->user()->can('delete', \App\Models\Reward::class))
+                                                        <th>Hành động</th>
+                                                    @endif
                                                 </tr>
                                             </thead>
                                             <tbody>
@@ -157,42 +159,46 @@
                                                                     style="width:80px; height:100px; object-fit:cover">
                                                             @endif
                                                         </td>
-                                                        <td>
-                                                            <button
-                                                                class="toggle-active-btn btn btn-xs {{ $reward->active == 1 ? 'btn-success' : 'btn-danger' }} text-white"
-                                                                data-id="{{ $reward->id }}"
-                                                                data-status="{{ $reward->active }}"
-                                                                data-url="{{ route('admin.rewards.changeActive') }}">
-                                                                {{ $reward->active == 1 ? 'Hiển thị' : 'Ẩn' }}
-                                                            </button>
+                                                        <td class="text-center">
+                                                            @can('changeActive', App\Models\Reward::class)
+                                                                <button
+                                                                    class="toggle-active-btn btn btn-xs {{ $reward->active == 1 ? 'btn-success' : 'btn-danger' }} text-white"
+                                                                    data-id="{{ $reward->id }}"
+                                                                    data-status="{{ $reward->active }}"
+                                                                    data-url="{{ route('admin.rewards.changeActive') }}">
+                                                                    {{ $reward->active == 1 ? 'Hiển thị' : 'Ẩn' }}
+                                                                </button>
+                                                            @else
+                                                                <span
+                                                                    class="badge light badge-{{ $reward->active == 1 ? 'success' : 'danger' }}">{{ $reward->active == 1 ? 'Hiển thị' : 'Ẩn' }}</span>
+                                                            @endcan
                                                         </td>
-
-
-
-                                                        <td>
-                                                            <div
-                                                                style="padding-right: 20px; display: flex; justify-content: end">
-                                                                @can('update', \App\Models\Reward::class)
-                                                                    <a href="{{ route('admin.rewards.edit', $reward->id) }}"
-                                                                        class="btn btn-primary shadow btn-xs sharp me-1">
-                                                                        <i class="fa fa-pencil"></i>
-                                                                    </a>
-                                                                @endcan
-                                                                @can('delete', \App\Models\Reward::class)
-                                                                    <form
-                                                                        action="{{ route('admin.rewards.delete', $reward->id) }}"
-                                                                        class="formDelete" method="post">
-                                                                        @csrf
-                                                                        @method('DELETE')
-                                                                        <button
-                                                                            class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
-                                                                            data-type="DELETE" href="">
-                                                                            <i class="fa fa-trash"></i>
-                                                                        </button>
-                                                                    </form>
-                                                                @endcan
-                                                            </div>
-                                                        </td>
+                                                        @if (auth()->user()->can('update', \App\Models\Reward::class) || auth()->user()->can('delete', \App\Models\Reward::class))
+                                                            <td>
+                                                                <div
+                                                                    style="padding-right: 20px; display: flex; justify-content: end">
+                                                                    @can('update', \App\Models\Reward::class)
+                                                                        <a href="{{ route('admin.rewards.edit', $reward->id) }}"
+                                                                            class="btn btn-primary shadow btn-xs sharp me-1">
+                                                                            <i class="fa fa-pencil"></i>
+                                                                        </a>
+                                                                    @endcan
+                                                                    @can('delete', \App\Models\Reward::class)
+                                                                        <form
+                                                                            action="{{ route('admin.rewards.delete', $reward->id) }}"
+                                                                            class="formDelete" method="post">
+                                                                            @csrf
+                                                                            @method('DELETE')
+                                                                            <button
+                                                                                class="btn btn-danger shadow btn-xs sharp me-1 call-ajax btn-remove btnDelete"
+                                                                                data-type="DELETE" href="">
+                                                                                <i class="fa fa-trash"></i>
+                                                                            </button>
+                                                                        </form>
+                                                                    @endcan
+                                                                </div>
+                                                            </td>
+                                                        @endif
                                                     </tr>
                                                 @endforeach
                                             </tbody>
@@ -207,7 +213,7 @@
                                     <div class="d-flex justify-content-center align-items-center p-5">
                                         <div>
                                             <h3 class="text-center">
-                                                {{ request()->name ? 'Chưa tìm thấy:' . request()->name : 'Không có dữ liệu' }}
+                                                {{ request()->name ? 'Không có kết quả với từ khóa:' . request()->name : 'Không có dữ liệu' }}
                                             </h3>
                                         </div>
                                     </div>
